requirements/base.txt
Django>=5.2,<5.3
gunicorn
psycopg2-binary
django-modeltranslation
django-mptt
django-autoslug
unidecode